Add config.in and initramfs.in for linux guests
Bareflank must be given a single ELF image to launch a linux guest.  To
this end, provide a mechanism for specifying how to embed an initramfs
into the final kernel image.

config/linux/tiny.config.in is a checked-in version of the output of
make tinyconfig from the linux tree, but with 64-bit and serial/tty
enabled. It sets CONFIG_INITRAMFS_SOURCE to @LINUX_INITRAMFS_SOURCE@.
On a call to configure_file, CMake will substitute @...@ with the value
of the config variable LINUX_INITRAMFS_SOURCE. This should be set by the
user to one of two things:

  1) a full path to an existing initramfs image or
  2) a full path to an "initramfs list" file (see below)

config/linux/tiny.initramfs.in is in the format expected by
usr/gen_init_cpio from the linux build tree, and provides a minimal
example of a root filesystem with a custom "hello" init binary.

The user can provide any initramfs list file they want through the
LINUX_INITRAMFS_IN variable.

Signed-off-by: Connor Davis <davisc@ainfosec.com>